--- slither/evm/__init__.py
@@ -1,2 +1,0 @@
-from .convert import SourceToEVM
-from .evm_cfg_builder import load_evm_cfg_builder
